Implement git-credential-osxkeychain workaround
On Mac OS X, if we detect "Apple Git" is installed and the user has
configured the eraseosxkeychain value, whenever we are asked to "store"
credentials that we ourselves emitted (because it was one of the AAD
or MSA authorities), we turn around and ask git-credential-osxkeychain
to "erase" what it stored _right before Git called us to store_,
ensuring that the hardcoded osxkeychain will never answer with
credentials that we manage.

This avoids the situation where the token we stored for an AAD or MSA
account is no longer valid (because it has expired or has been revoked),
and osxkeychain returns it anyway to Git (because it was hardcoded to
answer first) and Git promptly aborts the operation because of invalid
credentials, whereas we will validate a token before returning it to
Git, and, if we find that it has expired, we give the user the chance
to re-authenticate and acquire a new, valid token before returning
the token to Git.