TDD: parse_singleSquareBracketString
Azure Active Directory can return what appears to be
an array of integers and this would cause the parser to throw
an error as '[' was not a valid value character.  This is worked
around by considering square brackets as a fancy means of quoting
and as long as the array does not contain a nested array or strings
containing square brackets, we should be able to parse this JSON
enough to collect the "conformant" name/values.