Support Java 9+
There was a change made to the oauth2-useragent dependency to fix an
issue with parsing newer versions of Java. This change updates to that
newer version to include the fix.

Additionally, it appears javax.xml.bind is no longer included with Java
9. This has been included as a maven dependency to resolve class def
errors once the oauth2-useragent error was resolved.